Make window's pinned state a bit more correct
'Pinned' and 'sticky' are two distinct things: the former means that the
window is displayed on all workspaces, and the latter means that the
window should be kept in the same place on the "monitor glass" if the
viewport is scrolled.  (Not even sure what WMs even do viewports larger
than the monitor size anymore.)

Confusingly, 'sticky' seems to also get set when pinning a window?

At any rate, this stops using 'sticky' from Wnck as our version of
'pinned'.