Fix memory management of Wnck objects
Contrary to what was said in b379bcc7, we can't do without keeping a
reference on Wnck objects, because we can't control the references kept
on Xfw objects. We can therefore find ourselves in the situation where
an Xfw object outlives its Wnck reference object.

Moreover, keeping a reference on an object for the time of its use is a
normal thing to do, which we cannot blame the user for doing (whatever
the Libwnck documentation says).