XfwApplicationX11: Use WnckClassGroup instead of WnckApplication
What Libwnck calls "application" corresponds in fact to the instance of
an application (a single pid), while "class group" corresponds roughly
to the notion of application, as it can be inferred on Wayland from the
app-id.

Therefore what has been done in 702eafd3 does not allow to find a
satisfactory result for window grouping by application on X11, while the
result on Wayland is correct.

So let's deviate from Wnck naming on this point by starting to use
WnckClassGroup as a reference Wnck object for XfwApplication. This is
essentially a renaming since the two APIs are very similar.

The following commits continue this refactoring by introducing the
notion of instance as a member of XfwApplication, in order to retrieve
the pid, as well as the name and the list of windows per instance.