XfwScreenWayland: Properly update active window
Since we have no guarantee of the order in which the compositor sends
the change of window state (newly activated window first, deactivated
window next, or vice versa), we must emit
XfwScreenWayland::active-window-changed when a window loses its active
state if XfwScreenWayland:active-window has not already been updated.
So, in particular, when the stack of displayed windows empties,
XfwScreenWayland:active-window is properly reset.