Eliminate blurriness in X11 window icons
We need to tell libwnck the default icon size we want, otherwise it'll
give us icons of its default size, which are quite small.  Since we
don't want to keep invalidating the entirety of libwnck's icon cache any
time an app wants an icon of a different size, we just track the largest
icon size ever requested, and tell libwnck to generate icons of that
size.  If the caller requests something smaller, we downscale.

This isn't perfect, but is probably the best we can do.