Add xfw_(window|application)_get_gicon()
This also stops using libwnck's get_icon() and get_mini_icon() on X11,
as they're awkward to use and inflexible.  Instead, we can just fetch
_NET_WM_ICON ourselves, and get access to (possibly) several icons of
different sizes.

This also allows the GIcon to function properly with UI scaling, as for
some reason it seems GdkPixbuf (when used as a GIcon) doesn't support UI
scaling properly.