wayland: Fix monitor ownership
When first creating the XfwMonitorWayland instance, it should be moved
into the hash table, not copied.  This, way, if the screen gets shut
down before the first wl_output.done event, the monitor won't get
leaked.

When we eventually copy it into the screen's list of monitors, then it
should have a ref added, so there are three owners: the screen's monitor
list, the wl_output hash table, and the xdg_output hash table, each of
which will free on removal or destruction.