Use cosmic-workspace-unstable-v1 instead of ext-workspace
Labwc has chosen to vendor and implement Cosmic's workspace protocol.
It's very similar to the current ext-workspace draft, except that it's
behind on some of the changes that have been made more recently, and
still does the following "outdated" things:

* workspace event on the group instead of manager
* workspace_enter and workspace_leave on the group are missing
* wl_array used for capabilities and states rather than a bitfield

I've decided to maintain the existing semantics of our public API that
match the semantics of the current ext-workspace draft, as those
semantics were *so* much easier to deal with in applications than the
old way of doing things.