Add XfwWindowListMenu widget
This is mostly taken from the xfdesktop code, but the styling is more
like the xfce4-panel window-menu plugin.  It also includes the union of
the features supported by both.

The menu currently does not auto-update in an efficient way.  Every time
the menu is re-shown, it is re-populated.  The only exception is that if
a window is destroyed while the menu is being shown, the menu item for
it will be removed.

Closes #20