Use wnck_class_group_get_id() to attempt .desktop file icon fallback
The wnck_class_group_get_name() function uses a heuristic to attempt to
generate a friendly/display name for the class group, which may not
actually be the WM_CLASS resource name.  Since we do actually always
want the WM_CLASS resource name when trying to find the .desktop file
for the application, we need to use _get_id() instead.

Closes #33