meson-build: Use meson built-in tests
So that they are all at least compiled during meson dist in CI, and run
for those that can be non-interactively, even when this is of little
interest (we at least test that they don't crash).

Incidentally, you can also simply run meson test [--setup=full] to
launch the tests.