XfwMonitorWayland: Fix release order in _global_removed()
When the output was removed by the server before sending its properties,
and therefore before the monitor had been added to XfwScreen's list of
monitors (retaining a reference), the monitor was freed as soon as the
`g_hash_table_iter_remove(&iter)` instruction was issued, so
`monitor->output` was freed in `_finalize()`, and there was therefore a
double free in `g_clear_pointer(&monitor->output, ...`.

So `g_clear_pointer(&monitor->output, ...` must be done before
`g_hash_table_iter_remove(&iter)`, but this change also reorders the set
of operations in `_global_removed()`, so that they are performed in
reverse order to the reference additions.

Fixes: #49